PycharmProjects/byonic_phase1/requirements.txt
redis==3.5.3
SQLAlchemy==1.3.22
pandas
Flask==1.0.2
PyMySQL==0.9.3
Flask_Cors==3.0.6
PyJWT==2.0.1
python_dateutil==2.8.1

Werkzeug~=1.0.1
schedule~=1.0.0
python-dateutil~=2.8.1